+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.6 (unreleased)
 
 - Switched to mini-batch k-means
+- Improved performance of index creation for Postgres < 12
 
 ## 0.2.5 (2022-02-11)